feat: add option to configure OIDC prompt parameter (#14)
This introduces a new `oidc.prompt` configuration option to the Google
OIDC auth middleware. This allows for more control over the
authentication flow by specifying a prompt value, such as
`select_account`, to force the user to select which Google account to
authenticate with.

The integration tests have been enhanced to verify this new
functionality, and the `cookie.path` is now set in the test environment
to better scope cookies to their respective services. The README has
been updated to include documentation for the new option.

Close #12